- Simplified random fight number generation - Removed unnecessary variables
